Changes to allow stealing vehicles
Allow players to steal ped cars, they will not be locked out. However, after stealing the car the player is required to either hotwire or lockpick the vehicle to gain keys.

Stealing a parked car will have the player smash the window, which will give a chance to PoliceCall. Players are then required to hotwire or lockpick.
Lockpicking parked car doors will not trigger a PoliceCall.

Lastly, this uses qb-lock for the lockpick game. Hotwire however still has the progressbar.